// src/app/entities/models/CompanyBranches.ts
import { Common } from '../entities/Common';
export class CompanyBranches extends Common {
    PHONE: string;
    CODE: string;
    FAX: string;
    EMAIL: string;
    ADDRESS: string;
    ADDRESS2: string;
    ST_COM_ID: number;
    ST_CTY_ID: number;
    ST_CNT_ID: number;
    ST_CUR_CODE: string;
}